args global, readbility, rough loop, fixes
argc and argv[] have been made global, by declaring them outside of the
main function. How I didn't realise this was possible already is
irritating.

I changed the formatting of the error printf() call, so that it was a
little easier to read. I've also removed some unnecessary guff that was
lying around.

I have started to create the main loop, at the code to find the highest
prime number found. Now that I write this, I realise this is ridiculous
way to go about doing what I want.. I'll change it.

I also fixed the Fibonacci-generating loop, which was screwed in a
number of ways.