readme written, gitignore modified, llvm binary
I have written a load of explanation, instructions and background
information in the readme file. There's well cool tables and all sorts.
I think I've cleaned out any errors there may have been, and have tested
converting it to HTML with rdiscount.

I have also updated the .gitignore file, so that the Linux executable
should no longer be tracked. This appears not to have had an effect, so
the syntax may not be as I thought. The executable has been recompiled
with LLVM, however, and is a good deal smaller as a result.